@@ -15,32 +15,10 @@ import { rangesFromStringDefinition } from './workspace-util';
 
 export const EDITOR_MAX_LETTER = 1024;
 
-/**
- * Highlight a matching review comment with decorations (underline, icon next to the line)
- *
- * @param context The extensions context
- * @param selections The selection to highlight
- * @param editor The editor to work on
- * @return all highlighting decorations
- */
-export const gutterDecorations = (
-  context: ExtensionContext,
-  csvEntries: CsvEntry[],
-  editor: TextEditor,
-): TextEditorDecorationType[] => {
-  return [underlineDecoration(csvEntries, editor), commentIconDecoration(context, csvEntries, editor)];
-};
+export class Decorations {
+  constructor(private readonly context: ExtensionContext) {}
 
-/**
- * Highlight a matching review comment with decorations an underline decoration
- *
- * @param context The extensions context
- * @param selections The selection to highlight
- * @param editor The editor to work on
- * @return all highlighting decorations
- */
-export const underlineDecoration = (csvEntries: CsvEntry[], editor: TextEditor): TextEditorDecorationType => {
-  const decoration = window.createTextEditorDecorationType({
+  readonly decorationDeclarationType = window.createTextEditorDecorationType({
     isWholeLine: false,
     opacity: '0.9',
     borderWidth: '1px',
@@ -51,62 +29,70 @@ export const underlineDecoration = (csvEntries: CsvEntry[], editor: TextEditor):
     },
   });
 
-  const decorationOptions: DecorationOptions[] = [];
-
-  // build decoration options for each comment block
-  csvEntries.forEach((entry) => {
-    // iterate over multi-selections
-    rangesFromStringDefinition(entry.lines).forEach((range: Range) => {
-      decorationOptions.push({ range });
-    });
-  });
-  editor.setDecorations(decoration, decorationOptions);
-  return decoration;
-};
-
-/**
- * Highlight a matching review comment with an icon next to the start of the selection on the right
- *
- * @param context The extensions context
- * @param selections The selection to highlight
- * @param editor The editor to work on
- * @return all highlighting decorations
- */
-export const commentIconDecoration = (
-  context: ExtensionContext,
-  csvEntries: CsvEntry[],
-  editor: TextEditor,
-): TextEditorDecorationType => {
-  const decoration = window.createTextEditorDecorationType({
+  readonly commentDecorationType = window.createTextEditorDecorationType({
     isWholeLine: true,
     after: {
-      contentIconPath: context.asAbsolutePath(path.join('dist', 'speech-bubble-light.svg')),
+      contentIconPath: this.context.asAbsolutePath(path.join('dist', 'speech-bubble-light.svg')),
       margin: '5px',
     },
     dark: {
       after: {
-        contentIconPath: context.asAbsolutePath(path.join('dist', 'speech-bubble-dark.svg')),
+        contentIconPath: this.context.asAbsolutePath(path.join('dist', 'speech-bubble-dark.svg')),
       },
     },
   });
 
-  const decorationOptions: DecorationOptions[] = [];
+  /**
+   * Highlight a matching review comment with decorations an underline decoration
+   *
+   * @param csvEntries The selection to highlight
+   * @param editor The editor to work on
+   * @return all highlighting decorations
+   */
+  underlineDecoration(csvEntries: CsvEntry[], editor: TextEditor): void {
+    const decorationOptions: DecorationOptions[] = [];
 
-  // build decoration options for each comment block
-  csvEntries.forEach((entry) => {
-    // iterate over multi-selections
-    rangesFromStringDefinition(entry.lines).forEach((range: Range) => {
-      decorationOptions.push({
-        range: new Range(
-          new Position(range.start.line, EDITOR_MAX_LETTER),
-          new Position(range.start.line, EDITOR_MAX_LETTER),
-        ),
+    // build decoration options for each comment block
+    csvEntries.forEach((entry) => {
+      // iterate over multi-selections
+      rangesFromStringDefinition(entry.lines).forEach((range: Range) => {
+        decorationOptions.push({ range });
       });
     });
-  });
-  editor.setDecorations(decoration, decorationOptions);
-  return decoration;
-};
+
+    editor.setDecorations(this.decorationDeclarationType, decorationOptions);
+  }
+
+  /**
+   * Highlight a matching review comment with an icon next to the start of the selection on the right
+   *
+   * @param csvEntries The selection to highlight
+   * @param editor The editor to work on
+   * @return all highlighting decorations
+   */
+  commentIconDecoration(csvEntries: CsvEntry[], editor: TextEditor): void {
+    const decorationOptions: DecorationOptions[] = [];
+
+    // build decoration options for each comment block
+    csvEntries.forEach((entry) => {
+      // iterate over multi-selections
+      rangesFromStringDefinition(entry.lines).forEach((range: Range) => {
+        decorationOptions.push({
+          range: new Range(
+            new Position(range.start.line, EDITOR_MAX_LETTER),
+            new Position(range.start.line, EDITOR_MAX_LETTER),
+          ),
+        });
+      });
+    });
+    editor.setDecorations(this.commentDecorationType, decorationOptions);
+  }
+
+  clear(editor: TextEditor) {
+    this.underlineDecoration([], editor);
+    this.commentIconDecoration([], editor);
+  }
+}
 
 /**
  * Highlight a selection in an editor
